pyproject.toml
[project]
name = "quant-btc"
version = "0.1.0"
description = "BTC量化交易系统 — 符号主义规则引擎"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "matplotlib>=3.10.9",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "polars>=1.40.1",
    "pyarrow>=24.0.0",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
    "tzdata>=2026.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
